Add HashMap and set functions.
The first function just creates the bucket. The second function is the hash(key) function. It turns the key into a hash used as index of the bucket array. Then the set(key,value) adds a new element. If the key is the same the node gets overwrited, if they are different a new node is created